op-challenger: Remove remnants of fault.Player
